CodeBlockOverFlowDisposal/requirements_ocr.txt
PyMuPDF>=1.23.0
Pillow>=9.0.0
pytesseract>=0.3.10
numpy>=1.24.0